rafiq_project/requirements.txt
asgiref==3.11.0
boto3==1.42.1
botocore==1.42.1
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.9
django-storages==1.14.6
djangorestframework==3.16.1
filelock==3.20.3
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
pandas==3.0.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
regex==2026.1.15
requests==2.32.5
s3transfer==0.16.0
scikit-learn==1.6.1
scipy==1.16.3
six==1.17.0
sqlparse==0.5.4
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
